Customer List Link works to customer profile page
